dashboard/requirements.txt
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
requests==2.32.3
python-dotenv==1.0.1
streamlit-autorefresh==1.0.1
